Toolbox integrado en vista principal (falta alinear divs)
