+++ src/App.css
@@ -1,7 +1,6 @@
 .tab {
   text-align: center;
-  margin-top: 100px;
-  margin-left: 35%;
+
 }
 
 .App {
