--- src/controllers/Fabrica.css
@@ -1,20 +1,0 @@
-.celda {
-    width: 58px;
-    height: 58px;
-    border: 1px solid black;
-    background-color: #e6ecec;
-    display: inline-table;
-    margin-top: -4px;
-
-}
-
-.celda:hover {
-    background-color: antiquewhite;
-    border: 1px solid blue;
-}
-
-.tablero {
-    width: 600px;
-    height: 600px;
-
-}
