+++ src/controllers/Toolbox.css
@@ -2,8 +2,9 @@
     border: 1px solid slateblue;
     border-radius: 4px;
     background-color: darkblue;
-    height: 500px;
+    height: 600px;
     width: 140px;
+    display: inline-block;
 }
 
 .title{
